refactor(config): optimize config imports and function calls
This commit refactors the way the 'luminate' plugin handles its
configuration. Previously, the configuration was repeatedly imported
and accessed in multiple places, leading to verbose and repetitive
code.

Now, the configuration is imported once at the top of each file and
stored in a local variable. This makes the code cleaner and more
efficient, as the configuration is only imported once.

Additionally, the setup function in 'init.lua' has been refactored to
improve readability. The code for setting highlight groups and keymaps
has been moved into separate functions, making the setup function
easier to read and understand.

This refactor should not affect the functionality of the plugin, but
it makes the codebase easier to maintain and extend in the future.