feat: Add hlgroup option for yank, paste, undo, and redo
This commit introduces a new configuration option, hlgroup, for the
yank, paste, undo, and redo features in the luminate plugin. This
allows users to specify a custom highlight group for each of these
actions, providing greater flexibility in configuring the visual
appearance of these highlights.

The hlgroup option has been added to the README documentation, and
the corresponding code changes have been made in the luminate
configuration and autocmds files. Unnecessary code for removing
hlgroup in the setup function has been removed.

This change was made to address user requests for more customization
options and to improve the overall user experience with the luminate
plugin.